% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemeindemutationen.R
\name{gemeindemutationen}
\alias{gemeindemutationen}
\title{gemeindemutationen}
\usage{
gemeindemutationen(datum.neu, datum.daten = NA)
}
\arguments{
\item{datum.neu}{Datum, bis zu dem Mutationen im engeren Sinn berücksichtigt werden sollen. Das Datum muss im Format tt.mm.JJJJ eingegeben werden.}

\item{datum.daten}{Ein optionales Argument für das Datum, ab dem Mutationen im engeren Sinn berücksichtigt werden sollen. Das Datum muss im Format tt.mm.JJJJ eingegeben werden. Default ist NA. In diesem Fall werden alle Mutationen seit dem 01.01.1960 berücksichtigt.}
}
\value{
Ein Dataframe mit allen Mutationsmeldungen im engeren Sinn ab einem gewissen Zeitpunkt (frühestens 01.01.1960) und bis zu einem gewissen Zeitpunkt (spätestens 01.01.2020).
}
\description{
Mit der Funktion \code{gemeindemutationen} werden alle Mutationen im engeren Sinn, die Auswirkungen auf die Gemeindenummer haben, für einen gewünschten Zeitraum angezeigt.
}
\details{
\code{datum.neu} muss jünger sein als \code{datum.daten}, ansonsten wird ein leerer Dataframe zurückgegeben.

    Berücksichtigt werden können grundsätzlich alle Mutationen im engeren Sinn zwischen 01.01.1960 und 01.01.2020.

    Als Mutationen im engeren Sinn werden alle Gemeindefusionen, Gebietsänderungen, Eingemeindungen, Ausgemeindungen, Kantonswechsel, Bezirkswechsel und Neunummerierungen ausgewiesen, die mit einer Gebietsänderung und/oder einer Änderung der Gemeindenummer einher gehen.

    Die Daten stammen grundsätzlich vom Bundesamt für Statistik (Gemeindeverzeichnis), wurden aber für die vorliegenden Zwecke aufgearbeitet, erweitert und angepasst.
}
\examples{
\dontrun{

   gemeindemutationen(datum.daten = "01.01.1986", datum.neu = "01.01.1992")

   }

}
