% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munizipalgemeinde.R
\name{munizipalgemeinde}
\alias{munizipalgemeinde}
\title{munizipalgemeinde}
\usage{
munizipalgemeinde()
}
\value{
Ein Dataframe mit allen Recodes von Munizipal- auf politische Gemeinden.
}
\description{
Mit der Funktion \code{munizipalgemeinde} werden die im Kanton Thurgau vollzogenen Reorganisationen von Munizipal- zu politischen Gemeinden angezeigt.
}
\details{
Ab 1995 wurden im Kanton Thurgau die aus mehreren Ortsgemeinden bestehenden Munizipalgemeinden in politische Gemeinden überführt. Dies ging mit einer Änderung der Gemeindenummer einher. Mit der vorliegenden Funktion können diese Wechsel in der Nummerierung nachvollzogen werden.

    Die Daten stammen aus Vergleichen ausgewiesener Abstimmungsresultate auf Gemeindeebene mit den Thurgauer Gemeindenummern. Sie wurden für die vorliegenden Zwecke aufgearbeitet, erweitert und angepasst.
}
\examples{
\dontrun{

   munizipalgemeinde()

   }

}
