% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduzieren.R
\name{reduzieren}
\alias{reduzieren}
\title{reduzieren}
\usage{
reduzieren(
  datum.neu,
  datum.daten = NA,
  gemeinsame.urne = T,
  munizipalgemeinde = T
)
}
\arguments{
\item{datum.neu}{Datum, bis zu dem Mutationen berücksichtigt werden sollen. Das Datum muss im Format tt.mm.JJJJ eingegeben werden.}

\item{datum.daten}{Ein optionales Argument für das Datum, ab dem Mutationen berücksichtigt werden sollen. Das Datum muss im Format tt.mm.JJJJ eingegeben werden. Default ist NA. In diesem Fall werden alle Mutationen seit dem 01.01.1960 berücksichtigt.}

\item{gemeinsame.urne}{Ein optionales Argument. Es gibt an, ob gemeinsame Urnen angezeigt werden sollen oder nicht.}

\item{munizipalgemeinde}{Ein optionales Argument. Es gibt an, ob die Reorganisation von Munizipal- zu politischen Gemeinden im Kanton Thurgau angezeigt werden sollen oder nicht.}
}
\value{
Ein Dataframe mit allen Mutationsmeldungen ab einem gewissen Zeitpunkt (frühestens 01.01.1960) und bis zu einem gewissen Zeitpunkt (spätestens 01.01.2020).
}
\description{
Mit der Funktion \code{reduzieren} können alle Mutationsmeldungen im engeren Sinn, gemeinsame Urnen und Wechsel von Munizipal- zu politischen Gemeinden, die Auswirkungen auf die Gemeindenummern haben, für einen gewünschten Zeitraum angezeigt werden.
}
\details{
\code{datum.neu} muss jünger sein als \code{datum.daten}, ansonsten wird ein leerer Dataframe zurückgegeben.

    Berücksichtigt werden können grundsätzlich alle Mutationen zwischen 01.01.1960 und 01.01.2020.

    Als Mutationen werden alle Ausgemeindungen, Bezirkswechsel, Eingemeindungen, Gebietsänderungen, Gemeindefusionen, gemeinsame Urnen, Kantonswechsel, Neunummerierungen und Reorganisationen von Munizipal- zu politischen Gemeinden (Kanton Thurgau) ausgewiesen, die mit einer Gebietsänderung und/oder einer Änderung der Gemeindenummer einher gehen.

    Die Daten stammen von verschiedenen Quellen des Bundesamts für Statistik (Gemeindeverzeichnis, Abstimmungsresultate), wurden aber für die vorliegenden Zwecke aufgearbeitet, erweitert und angepasst.
}
\examples{
\dontrun{

   reduzieren(datum.daten = "01.01.1986", datum.neu = "01.01.1992", gemeinsame.urne=T,
   munizipalgemeinde=F)

   }

}
